requirements.txt
# =========================
# PARSER / DOCUMENTOS
# =========================
docling>=2.86.0,<3.0.0
pymupdf>=1.27.2.2,<2.0.0
pytesseract>=0.3.13,<0.4.0
pdf2image>=1.16.3,<2.0.0
pdfplumber>=0.10.0,<1.0.0
python-docx>=1.1.0,<2.0.0
rarfile>=4.2,<5.0
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
xlrd>=2.0.0,<3.0.0

# =========================
# ORQUESTRAÇÃO / RAG
# =========================
langgraph>=1.1.6,<2.0.0
llama-index>=0.14.20,<1.0.0
langsmith>=0.7.30,<1.0.0

# =========================
# LLM / EMBEDDINGS
# =========================
anthropic>=0.94.0,<1.0.0
sentence-transformers>=5.4.0,<6.0.0
transformers>=5.5.3,<6.0.0

# =========================
# BUSCA / RETRIEVAL
# =========================
qdrant-client>=1.17.1,<2.0.0
bm25s>=0.2.0                   # preferred BM25 library (Phase 2)
rank-bm25>=0.2.2,<1.0.0        # fallback BM25 — used directly in bm25_retriever.py
nltk>=3.8                      # Portuguese stopwords tokenizer
spacy>=3.7                     # optional: richer PT NLP pipeline
pydantic>=2.0                  # RetrievalResult and future Pydantic models

# =========================
# API / BACKEND
# =========================
fastapi>=0.135.3,<1.0.0
uvicorn>=0.44.0,<1.0.0
python-dotenv>=1.2.2,<2.0.0

# =========================
# FRONTEND
# =========================
streamlit>=1.56.0,<2.0.0

# =========================
# AVALIAÇÃO / TESTES
# =========================
ragas>=0.4.3,<1.0.0
pytest>=9.0.3,<10.0.0

# =========================
# UTILITÁRIOS
# =========================
curl_cffi>=0.15.0,<1.0.0
tqdm>=4.66.0,<5.0.0
tenacity>=9.1.4,<10.0.0

# =========================
# FILE SYSTEM / ARQUIVOS
# =========================
filetype>=1.2.0,<2.0.0

# =========================
# FASE 2 — NLTK / SPACY DATA
# =========================
# After installing, run once:
#   python -m nltk.downloader stopwords punkt
#   python -m spacy download pt_core_news_sm
